+++ sale_invoice_plan/models/__init__.py
@@ -1,4 +1,5 @@
 # Copyright 2019 Ecosoft Co., Ltd (http://ecosoft.co.th/)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
+from . import res_config_settings
 from . import sale
 from . import sale_invoice_plan